pyproject.toml
[tool.poetry]
name = "Iridium"
version = "1.1"
description = "A data conversion and processing platform."
authors = ["Raghavendra Pratap Singh <raghavendra_ps@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# Web Framework & Server
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
gunicorn = "^20.1.0"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"

# Database (SQLAlchemy ORM)
sqlalchemy = "^2.0.9"
psycopg2-binary = "^2.9.6"
alembic = "^1.10.3" # For database migrations

# Async Task Queue
celery = "^5.2.7"
redis = "^4.5.4"

# Authentication & Security
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

# Configuration Management
pydantic = {extras = ["email"], version = "^1.10.7"}

# Data Processing Libraries (add more as needed)
pandas = "^2.0.0"
openpyxl = "^3.1.2" # For .xlsx support with pandas

# File Processing & OCR
pytesseract = "^0.3.10"
pillow = "^9.5.0"
pypdf2 = "^3.0.1"
python-docx = "^0.8.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
httpx = "^0.24.0" # For async testing of the API

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
